requirements.txt
flask
flask-cors
torch
torchvision
numpy
pillow
opencv-python
pandas
scikit-learn
tqdm
